Refactor code to make it more testable going forward
This is still a bit of a WIP but our index.ts file should allow for us to dictate how DB and server connections are defined. This is so that we can differentiate between when we want to test and when we want to run the server with real data and a real database

This commit helps us achieve that, albeit not in the optimal way just yet.