Refactor Location model on backend and frontend
Our Location model was using two properties, Lat and Lng, for coordinates.

It turns out that MongoDB has a built in type for geo coordinates which allows for database level aggregations and look ups related to locations.

Because of that we're shifting our Location model to using the MongoDB built in GeoSpatial type.

NOTE: MongoDB stores coordinates as LONGITUDE and LATITUDE, which is the opposite of what, eg, Google Maps, consumes (Latitude then Longitude). Keep this in mind when writing location based logic and storing things in the database, otherwise we could end up with our map pins placed on their opposite end of the planet ^^"